Implement multi-factor requirement

+++ wcfsetup/install/files/lib/form/MultifactorDisableForm.class.php
@@ -8,6 +8,7 @@ use wcf\page\AccountSecurityPage;
 use wcf\system\exception\IllegalLinkException;
 use wcf\system\exception\PermissionDeniedException;
 use wcf\system\form\builder\field\BooleanFormField;
+use wcf\system\form\builder\field\RejectEverythingFormField;
 use wcf\system\form\builder\field\validation\FormFieldValidationError;
 use wcf\system\form\builder\field\validation\FormFieldValidator;
 use wcf\system\form\builder\LanguageItemFormNode;
@@ -101,20 +102,34 @@ class MultifactorDisableForm extends AbstractFormBuilderForm
     {
         parent::createForm();
 
+        $remaining = $this->setupsWithoutDisableRequest(
+            $this->setupsWithoutBackupCodes($this->setups)
+        );
+
+        if (WCF::getUser()->requiresMultifactor() && empty($remaining)) {
+            $this->form->addDefaultButton(false);
+            $this->form->appendChildren([
+                LanguageItemFormNode::create('requireMultifactor')
+                    ->languageItem('wcf.user.security.multifactor.disable.requireMultifactor')
+                    ->variables([
+                        'setup' => $this->setup,
+                    ]),
+                RejectEverythingFormField::create(),
+            ]);
+
+            return;
+        }
+
         $this->form->appendChildren([
             LanguageItemFormNode::create('explanation')
                 ->languageItem('wcf.user.security.multifactor.disable.explanation')
                 ->variables([
-                    'remaining' => $this->setupsWithoutDisableRequest(
-                        $this->setupsWithoutBackupCodes($this->setups)
-                    ),
+                    'remaining' => $remaining,
                     'setup' => $this->setup,
                 ]),
             BooleanFormField::create('confirm')
                 ->label('wcf.user.security.multifactor.disable.confirm', [
-                    'remaining' => $this->setupsWithoutDisableRequest(
-                        $this->setupsWithoutBackupCodes($this->setups)
-                    ),
+                    'remaining' => $remaining,
                     'setup' => $this->setup,
                 ])
                 ->addValidator(new FormFieldValidator('confirm', static function (BooleanFormField $formField) {
@@ -148,6 +163,10 @@ class MultifactorDisableForm extends AbstractFormBuilderForm
         $remaining = $this->setupsWithoutBackupCodes($setups);
 
         if (empty($remaining)) {
+            if (WCF::getUser()->requiresMultifactor()) {
+                throw new \LogicException('The user requires multi-factor authentication.');
+            }
+
             foreach ($setups as $setup) {
                 $setup->delete();
             }
